split road table components into containers
